[TASK: Redoing font rendering] -> Corrected all warnings in the project. WE'RE CLEAN!

+++ Gene/Engine/Private/Graphics/Font.cc
@@ -47,7 +47,7 @@ Vector2 Font::MeasureString(const std::string & str)
 
         if (glyph->Height > height)
         {
-            height = glyph->Height;
+            height = static_cast<float>(glyph->Height);
         }
 
         i++;
